requirements.txt
# This file was autogenerated by uv via the following command:
#    ./scripts/generate_requirements.sh
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via httpx
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
cohere==5.15.0
    # via python-template (pyproject.toml)
fastavro==1.11.1
    # via cohere
filelock==3.18.0
    # via huggingface-hub
fsspec==2025.5.1
    # via huggingface-hub
grpcio==1.72.1
    # via qdrant-client
h11==0.16.0
    # via httpcore
h2==4.2.0
    # via httpx
hf-xet==1.1.3
    # via huggingface-hub
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   cohere
    #   langsmith
    #   qdrant-client
httpx-sse==0.4.0
    # via cohere
huggingface-hub==0.32.4
    # via tokenizers
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain-core==0.3.64
    # via langchain-text-splitters
langchain-text-splitters==0.3.8
    # via python-template (pyproject.toml)
langsmith==0.3.45
    # via langchain-core
neo4j==5.28.1
    # via python-template (pyproject.toml)
numpy==2.3.0
    # via qdrant-client
orjson==3.10.18
    # via langsmith
packaging==24.2
    # via
    #   huggingface-hub
    #   langchain-core
    #   langsmith
portalocker==2.10.1
    # via qdrant-client
protobuf==6.31.1
    # via qdrant-client
pydantic==2.11.5
    # via
    #   python-template (pyproject.toml)
    #   cohere
    #   langchain-core
    #   langsmith
    #   qdrant-client
pydantic-core==2.33.2
    # via
    #   cohere
    #   pydantic
python-dotenv==1.1.0
    # via python-template (pyproject.toml)
pytz==2025.2
    # via neo4j
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   langchain-core
qdrant-client==1.14.2
    # via python-template (pyproject.toml)
requests==2.32.3
    # via
    #   python-template (pyproject.toml)
    #   cohere
    #   huggingface-hub
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
sniffio==1.3.1
    # via anyio
tenacity==9.1.2
    # via langchain-core
tokenizers==0.21.1
    # via cohere
tqdm==4.67.1
    # via huggingface-hub
types-requests==2.32.0.20250602
    # via cohere
typing-extensions==4.14.0
    # via
    #   anyio
    #   cohere
    #   huggingface-hub
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.4.0
    # via
    #   qdrant-client
    #   requests
    #   types-requests
zstandard==0.23.0
    # via langsmith
